#63 - refactor: 댓글 Service 단일 책임 원칙 적용
